+++ tests/test_sql.py
@@ -42,7 +42,33 @@ class Job(SQLModel):
 
 class JobRole(SQLModel):
     name = Str().tag(length=64)
+    default = Bool()
     job = Instance(Job)
+
+    check_one_default = sa.schema.DDL('''
+        CREATE OR REPLACE FUNCTION check_one_default() RETURNS TRIGGER
+        LANGUAGE plpgsql
+        AS $$
+        BEGIN
+            IF EXISTS (SELECT * from test_sql.JobRole
+                       WHERE "default" == true AND "job" == NEW."job") THEN
+                RAISE EXCEPTION 'A default aleady exists';
+            END IF;
+            RETURN ENW;
+        END;
+        $$;''')
+
+    trigger = sa.schema.DDL('''
+        CREATE CONSTRAINT TRIGGER check_default_role AFTER INSERT OR UPDATE
+        OF "default" on ON test_sql.JobRole
+        FOR EACH ROW EXECUTE FUNCTION check_one_default();''')
+
+    class Meta:
+        triggers = {
+            'after_create':
+                lambda: JobRole.trigger.execute_if(dialect='postgresql')
+        }
+
 
 
 class ImageInfo(JSONModel):